lesson2/requirements.txt
flask 
flask-sqlalchemy
psycopg2-binary
PyJWT